+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## Added
 
+- Custom key bindings can be set.
+- Racings can be paused and restarted.
 - The character runs forwards and can strafe left and right.